[Feature #15] Remove all 'any' types for better type safety
- Removed ESLint disable comment for no-explicit-any rule
- Added proper TypeScript interfaces for NextAuth:
  - Created comprehensive type declarations in next-auth.d.ts
  - Properly typed JWT token with id property
  - Properly typed Session with user id
  - Added username property to User interface
- Updated auth-config.ts to use proper types:
  - Added AuthOptions type for authOptions export
  - Typed JWT callback parameters (token: JWT, user?: User)
  - Typed session callback parameters (session: Session, token: JWT)
  - Typed createUser event parameter (user: User)
- Fixed pages configuration (changed signUp to newUser per NextAuth spec)
- Consolidated type declarations into single next-auth.d.ts file at root
- All TypeScript errors resolved, no 'any' types remaining

Resolves #15
