+++ next-auth.d.ts
@@ -1,0 +1,24 @@
+import { DefaultSession, DefaultUser } from "next-auth"
+import { JWT as DefaultJWT } from "next-auth/jwt"
+
+declare module "next-auth" {
+  interface Session {
+    user: {
+      id: string
+      name?: string | null
+      email?: string | null
+      image?: string | null
+    } & DefaultSession["user"]
+  }
+
+  interface User extends DefaultUser {
+    id: string
+    username?: string
+  }
+}
+
+declare module "next-auth/jwt" {
+  interface JWT extends DefaultJWT {
+    id?: string
+  }
+}
